+++ CHANGELOG.md
@@ -2,6 +2,66 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.3.0](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.2.0...v1.3.0) (2024-05-07)
+
+
+### Features
+
+* support command override ([#284](https://github.com/aws-actions/amazon-ecs-render-task-definition/issues/284)) ([9259793](https://github.com/aws-actions/amazon-ecs-render-task-definition/commit/925979370e7456028a520e0dad54854d027a3808))
+
+## [1.2.0](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.1.3...v1.2.0) (2023-11-14)
+
+
+### Features
+
+* ability to configure log driver and log options ([#255](https://github.com/aws-actions/amazon-ecs-render-task-definition/issues/255)) ([c6f3dfb](https://github.com/aws-actions/amazon-ecs-render-task-definition/commit/c6f3dfbdd9990283e47c9e0cf46393ea13f1541b))
+* support dockerLabels ([#274](https://github.com/aws-actions/amazon-ecs-render-task-definition/issues/274)) ([b9718e5](https://github.com/aws-actions/amazon-ecs-render-task-definition/commit/b9718e5f28bbcaeae2944140d52452f71601a622))
+
+
+### Bug Fixes
+
+* update ncc and minor workflow enhancements ([#269](https://github.com/aws-actions/amazon-ecs-render-task-definition/issues/269)) ([3e9a743](https://github.com/aws-actions/amazon-ecs-render-task-definition/commit/3e9a74317f3b8a55dc9b95b57a1cac50099ed343))
+
+### [1.1.3](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.1.2...v1.1.3) (2023-01-17)
+
+### [1.1.2](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.1.1...v1.1.2) (2022-10-21)
+
+### [1.1.1](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.1.0...v1.1.1) (2022-01-18)
+
+
+### Bug Fixes
+
+* Strict Mode Deprecation ([#150](https://github.com/aws-actions/amazon-ecs-render-task-definition/issues/150)) ([1d693a5](https://github.com/aws-actions/amazon-ecs-render-task-definition/commit/1d693a567ca030686d891af532ada9b4d908f863))
+
+## [1.1.0](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.24...v1.1.0) (2021-12-01)
+
+
+### Features
+
+* add environment-variables support ([#114](https://github.com/aws-actions/amazon-ecs-render-task-definition/issues/114)) ([85f5e35](https://github.com/aws-actions/amazon-ecs-render-task-definition/commit/85f5e350c104582a36945a4b66abdd29b42123bf))
+
+### [1.0.24](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.23...v1.0.24) (2021-08-24)
+
+### [1.0.23](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.22...v1.0.23) (2021-07-06)
+
+### [1.0.22](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.21...v1.0.22) (2021-06-01)
+
+### [1.0.21](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.20...v1.0.21) (2021-05-10)
+
+### [1.0.20](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.19...v1.0.20) (2021-03-02)
+
+### [1.0.19](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.18...v1.0.19) (2021-02-02)
+
+### [1.0.18](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.17...v1.0.18) (2021-01-19)
+
+### [1.0.17](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.16...v1.0.17) (2020-11-11)
+
+### [1.0.16](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.15...v1.0.16) (2020-10-29)
+
+### [1.0.15](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.14...v1.0.15) (2020-10-05)
+
+### [1.0.14](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.13...v1.0.14) (2020-08-25)
+
 ### [1.0.13](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.12...v1.0.13) (2020-08-11)
 
 ### [1.0.12](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.0.11...v1.0.12) (2020-07-17)